move graph-tracking to utils
