+++ src/mygrad/nnet/losses/multiclass_hinge.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-import mygrad._graph_tracking as _tracking
+import mygrad._utils.graph_tracking as _tracking
 from mygrad.operation_base import Operation
 from mygrad.tensor_base import Tensor
