+++ src/mygrad/nnet/losses/softmax_crossentropy.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-import mygrad._graph_tracking as _tracking
+import mygrad._utils.graph_tracking as _tracking
 from mygrad.math._special import logsumexp
 from mygrad.operation_base import Operation
 from mygrad.tensor_base import Tensor
